refactor: query permissions within PermissionsList component
